move back NN extension
