+++ Project.toml
@@ -12,16 +12,15 @@ Dictionaries = "85a47980-9c8c-11e8-2b9f-f7ca1fa99fb4"
 Distances = "b4f34e82-e78d-54a5-968a-f98e89d6e8f7"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Meshes = "eacbb407-ea5a-433e-ab97-5258b1ca43fa"
+NearestNeighbors = "b8a86587-4115-5ab1-83bc-aa920d37bbce"
 PlotlyExtensionsHelper = "a7e57fc7-3548-4021-90c3-ac0e175068b1"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 
 [weakdeps]
-NearestNeighbors = "b8a86587-4115-5ab1-83bc-aa920d37bbce"
 PlotlyBase = "a03496cd-edff-5a9b-9e67-9cda94a718b5"
 
 [extensions]
-NearestNeighborsExt = "NearestNeighbors"
 PlotlyBaseExt = "PlotlyBase"
 
 [compat]
